+++ NEWS.md
@@ -6,6 +6,8 @@ Phan NEWS
 New features (Analysis):
 + Don't emit `PhanTypeInvalidLeftOperandOfBitwiseOp` and other binary operation warnings for `mixed`
 + Emit `PhanIncompatibleRealPropertyType` when real property types are incompatible (#4016)
++ Change the way `PhanIncompatibleCompositionProp` is checked for. (#4024)
+  Only emit it when the property was redeclared in an inherited trait.
 
 Plugins:
 + Warn and skip checks instead of crashing when running `InlineHTMLPlugin` without the `tokenizer` extension installed. (#3998)